update app to Laravel 10
