+++ app/Providers/AppServiceProvider.php
@@ -25,6 +25,5 @@ class AppServiceProvider extends ServiceProvider
      */
     public function boot()
     {
-        Paginator::useTailwind();
     }
 }
