+++ composer.json
@@ -8,28 +8,25 @@
     ],
     "license": "MIT",
     "require": {
-        "php": "^7.2.5",
-        "doctrine/dbal": "^2.10",
-        "fideloper/proxy": "^4.2",
-        "fruitcake/laravel-cors": "^1.0",
-        "guzzlehttp/guzzle": "^6.3",
-        "laravel/cashier": "^12.2",
-        "laravel/framework": "^7.0",
-        "laravel/tinker": "^2.0",
-        "laravel/ui": "^2.1",
-        "spatie/laravel-honeypot": "^2.2",
-        "spatie/laravel-newsletter": "^4.8",
-        "spatie/laravel-sluggable": "^2.2"
+        "php": "^8.1.0",
+        "doctrine/dbal": "^3.0",
+        "drewm/mailchimp-api": "^2.5",
+        "guzzlehttp/guzzle": "^7.0.1",
+        "laravel/framework": "^10.0",
+        "laravel/tinker": "^2.9.0",
+        "laravel/ui": "^4.4.0",
+        "spatie/laravel-honeypot": "^4.4.0",
+        "spatie/laravel-newsletter": "^5.1.1",
+        "spatie/laravel-sluggable": "^3.5.0"
     },
     "require-dev": {
         "barryvdh/laravel-debugbar": "^3.3",
-        "facade/ignition": "^2.0",
-        "fzaninotto/faker": "^1.9.1",
-        "jasonmccreary/laravel-test-assertions": "^1.0",
-        "laravel-shift/blueprint": "^1.15",
+        "spatie/laravel-ignition": "^2.0",
+        "jasonmccreary/laravel-test-assertions": "^2.3.0",
+        "laravel-shift/blueprint": "^2.9.0",
         "mockery/mockery": "^1.3.1",
-        "nunomaduro/collision": "^4.1",
-        "phpunit/phpunit": "^8.5"
+        "nunomaduro/collision": "^7.0",
+        "phpunit/phpunit": "^10.0"
     },
     "config": {
         "optimize-autoloader": true,
@@ -43,19 +40,17 @@
     },
     "autoload": {
         "psr-4": {
-            "App\\": "app/"
-        },
-        "classmap": [
-            "database/seeds",
-            "database/factories"
-        ]
+            "App\\": "app/",
+            "Database\\Factories\\": "database/factories/",
+            "Database\\Seeders\\": "database/seeders/"
+        }
     },
     "autoload-dev": {
         "psr-4": {
             "Tests\\": "tests/"
         }
     },
-    "minimum-stability": "dev",
+    "minimum-stability": "stable",
     "prefer-stable": true,
     "scripts": {
         "post-autoload-dump": [
